requirements.txt
discord.py>=2.3.2
PyYAML>=6.0.1
python-dateutil>=2.8.2 